Major Cleanup of Preferences
-Place Font Typeface & Weight in separate preferences
(So we do not have to parse the preferences, and it makes the code less
confusing.)

-Remove SystemUI and Android System fonts setting saved in Main Pref
and place it in respective preference
(My initial reason for placing SystemUI and Android System in the main
preference was to speed up preference reading because I didn’t want to
slow boot time by reading from a large file.
However, the speed difference is negligible and it just adds bulk to
the code.)

-Some code cleanups

- Changed preference file names
(because we “Place Font Typeface & Weight in separate preferences”, if
users do not clear data, it will force close on them. We want to be
safe and reduce confusion)